Option to clone without overwriting existing files

+++ src/cloneFile.js
@@ -1,5 +1,6 @@
 const fs = require('fs').promises;
 const path = require('path');
+const COPYFILE_EXCL = require('fs').constants.COPYFILE_EXCL;
 
 /**
  * Copy file or tree to a new location
@@ -7,10 +8,26 @@ const path = require('path');
  *
  * @param {string} sourcePath - old path
  * @param {string} destinationPath - new path
+ * @param {Object} [options] - options object (optional)
+ * @param {boolean} [options.overwrite=true] - will not overwrite existing file
  */
-const cloneFile = async function cloneFile(sourcePath, destinationPath) {
+const cloneFile = async function cloneFile(
+  sourcePath,
+  destinationPath,
+  options
+) {
   await fs.mkdir(path.dirname(destinationPath), { recursive: true });
-  return fs.copyFile(sourcePath, destinationPath);
+  if (options && !options.overwrite) {
+    return fs
+      .copyFile(sourcePath, destinationPath, COPYFILE_EXCL)
+      .catch(function (err) {
+        if (!/^EEXIST: file already exists, copyfile /.test(err.message)) {
+          throw err;
+        }
+      });
+  } else {
+    return fs.copyFile(sourcePath, destinationPath);
+  }
 };
 
 module.exports = cloneFile;
